delivery/pyproject.toml
[tool.poetry]
name = "courier-services"
description = "Globerce courier API services"
version = "1.0.0"
packages = [
    { include = 'api' },
    { include = 'cli' },
]

authors = []

[[tool.poetry.source]]
name = "globerce"
url = "https://python.globerce.com/repository/pypi-proxy/simple"
priority = "primary"


[[tool.poetry.source]]
name = "PyPI"
priority = "supplemental"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
aiohttp = "^3.8.1"
aioredis = "^2.0.0"
numpy = "1.22.0"
cacheout = "^0.14.1"
email-validator = "^1.1.3"
itsdangerous = "^2.0.1"
mode = "^4.3.2"
python-dateutil = "^2.8.2"
python-magic = "^0.4.24"
stringcase = "^1.2.0"
uvicorn = "0.15.0"
websockets = "^10.2"
python-multipart = "^0.0.5"
aiofiles = "^22.1.0"
pandas = "^1.4.2"
openpyxl = "^3.0.9"
XlsxWriter = "^3.0.3"
psycopg2-binary = "^2.9.3"
loguru = "^0.6.0"
pdfkit = "^1.0.0"
Pillow = "^9.1.0"
python-slugify = "^7.0.0"
ortools = "^9.3.10497"
Shapely = "^1.8.2"
transliterate = "^1.10.2"
aerich = "0.6.2"
asyncpg = "^0.27.0"
fastapi-pagination = "^0.11.2"
SQLAlchemy = "^1.4.45"
httpx = "^0.23.3"
pytest-asyncio = "^0.20.3"
tortoise-orm = "0.18.1"
fastapi = "^0.95.0"
fastapi-mail = "^1.2.7"
num2words = "^0.5.12"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
bcrypt = "4.0.1"
transitions = "0.9.2"

[tool.poetry.group.dev.dependencies]
async-asgi-testclient = "^1.4.11"
flake8-bugbear = "^22.12.6"
flake8-commas = "^2.1.0"
flake8-comprehensions = "^3.10.1"
flake8-isort = "^6.0.0"
flake8-quotes = "^3.3.2"
flake8 = "^6.0.0"
pytest-asyncio = "^0.20.3"
pytest_cov = "^3.0.0"
pytest-freezegun = "^0.4.2"
pytest-mock = "^3.10.0"
pytest-xdist = "^3.1.0"
pytest = "^7.2.0"

[tool.poetry.scripts]
delivery-api = 'api.__main__:main'
delivery-cli = 'cli.__main__:main'
